Separation of jvm-specific code from TableSchemaUpdateConnectionTest
